Add section loop for ab test. Add a disable option for playback settings. Implement 'playedOnce' and 'loopTimes' features for training examples. Fix a bug causing no value filled in CSV file for rating tests.
