+++ listening-test-react/src/views/components/RenderRatingExample.tsx
@@ -1,13 +1,12 @@
 import {observer} from "mobx-react";
 import {AudioButton, AudioController, useAudioPlayer} from "../../shared/web-audio/AudiosPlayer";
 import {AudioLoading, useAllAudioReady} from "../../shared/web-audio/AudiosLoading";
-import React, {RefObject, useEffect, useState} from "react";
+import React, {useEffect, useState} from "react";
 import Grid from "@material-ui/core/Grid";
 import {RenderSurveyControl} from "../../shared/components/RenderSurveyControl";
 import {ratingAreaStyle} from "../SharedStyles";
-import {Box, createStyles, Slider, Theme, Typography} from "@material-ui/core";
-import {makeStyles} from "@material-ui/core/styles";
 import {AudioExampleModel, AudioFileModel} from "../../shared/models/AudioTestModel";
+import {AudioSectionLoopingController} from "../../shared/web-audio/AudioSectionLoopingController";
 
 export const RenderRatingExample = observer(function (props: { value: AudioExampleModel, RatingBar: (props: { audio: AudioFileModel }) => JSX.Element, active?: boolean }) {
   const {value, RatingBar, active} = props;
@@ -47,51 +46,3 @@ export const RenderRatingExample = observer(function (props: { value: AudioExamp
     </Grid>
   </>
 })
-
-const useStyles = makeStyles((_: Theme) => createStyles({
-  valueLabel: {top: 22, '& *': {background: 'transparent', color: '#000'}},
-  thumb: {zIndex: 1},
-  barContainer: {position: 'relative'},
-  barCover: {position: 'absolute', top: -20, height: 2, display: 'block', background: '#b7bde3'}
-}));
-const AudioSectionLoopingController = function ({refs, currentTime, setTimeUpdate}: { refs: RefObject<HTMLAudioElement> [], currentTime: number, setTimeUpdate: (_: () => void) => void}) {
-  const primary = refs[0];
-  const [range, setRange] = useState([0, 0]);
-  const classes = useStyles();
-  useEffect(() => {
-    // Make sure it has value and set range
-    if (primary?.current?.duration) {
-      const newValue = [0, primary.current?.duration];
-      setRange(newValue);
-    }
-  }, [primary.current]);
-  // Return update methods for parent component
-  useEffect(() => setTimeUpdate(() => {
-    if (currentTime < range[0] || currentTime > range[1])
-      refs.forEach(value1 => value1.current.currentTime = range[0])
-  }));
-
-  // When drag the slider bar
-  const handleChange = (_: any, newValue: number | number[]) => {
-    const newRanges = newValue as number[]
-    // If the current time is out of range
-    if (currentTime < newRanges[0] || currentTime > newRanges[1])
-      refs.forEach(value1 => value1.current.currentTime = newRanges[0])
-    setRange(newRanges);
-  }
-  const handleSliderLabelFormat = (num: number) => {
-    return isNaN(num) ? 0 : num.toFixed(0) + 's'
-  }
-  // Calculate cover bar of styles every rendering
-  const barCoverStyle = {
-    left: 0, width: range[0] / refs[0].current?.duration * 100 + '%'
-  };
-
-  return <Box className={classes.barContainer}>
-    <Typography>Section Looping Controls</Typography>
-    <Slider classes={{thumb: classes.thumb, valueLabel: classes.valueLabel}} valueLabelDisplay="auto"
-            value={range} aria-labelledby="range-slider" step={0.1} max={refs[0].current?.duration}
-            valueLabelFormat={handleSliderLabelFormat} onChange={handleChange}/>
-    <span className={classes.barCover} style={barCoverStyle}/>
-  </Box>
-}
